Made unit tests for the Salt framework
The directory `tests` will consist of small applications using the Salt framework, that will contain unit tests for testing the framework.

This commit also features a fix to the Salt Loader. There was a bug, that if a directory in the absolute path features the same name as the main package of the application using the framework, the annotation processor would throw an exception while creating a new instance of the object because the class name would be parsed false.